Use `$marker_length` to correctly parse conflict markers
Since Jujutsu v0.25.0, conflict markers are now allowed to be longer
than 7 characters. The new `$marker_length` variable can be used to pass
the length of the markers to the merge tool.

In order to support longer conflict markers, the `merge-args` setting
for jj-diffconflicts needs to be updated to pass the marker length:

    merge-args = [
      "-c", "let g:jj_diffconflicts_marker_length=$marker_length",
      "-c", "JJDiffConflicts!", "$output", "$base", "$left", "$right"
    ]

But unless long conflict markers are used, jj-diffconflicts should still
work without the use of `$marker_length`, or with Jujutsu versions older
than v0.25.0.

References:
- https://jj-vcs.github.io/jj/latest/conflicts/#long-conflict-markers
- https://jj-vcs.github.io/jj/v0.25.0/config/#setting-up-a-custom-merge-tool
- https://github.com/jj-vcs/jj/releases/tag/v0.25.0