Handle conflicts with missing terminating newline
Since v0.26.0, when Jujutsu encounters a missing terminating newline
character in a conflict, it will add a comment to the conflict markers.

The way we are handling this use case is by actually making the regular
expressions for matching conflict markers less strict: we now only check
that there are trailing comments, regardless of whether they are in a
format that we expect or not. In theory, this could have the potential
of matching lines that are not conflict markers, but I believe that this
is not likely now that Jujutsu uses long conflict markers to
disambiguate conflict markers from file content.

References:
- https://jj-vcs.github.io/jj/latest/conflicts/#conflicts-with-missing-terminating-newline
- https://jj-vcs.github.io/jj/latest/conflicts/#long-conflict-markers